// src/interfaces/operations-data-type.ts
export type OperationsCaliforniaData = {
    id: number;
    closing_date: string;
    week: number;// Week [Entero]
    roof: number; // Roof [Entero]
    water: number; // Water [Entero]
    mold: number; // Mold [Entero]
    mold_testing: number; // Mold testing [Entero]
    shrink_wrap: number; // Shrink Wrap [Entero]
    inspections: number; // Inspections [Entero]
    second_tarp: number; // 2nd Tarp [Entero]
    observations: string;
    created_at: string;
    updated_at: string;
}
export type OperationsMiamiData = {
    id: number;
    closing_date: string;
    week: number;// Week [Entero]
    roof: number; // Roof [Entero]
    water: number; // Water [Entero]
    mold: number; // Mold [Entero]
    mold_testing: number; // Mold testing [Entero]
    shrink_wrap: number; // Shrink Wrap [Entero]
    inspections: number; // Inspections [Entero]
    second_tarp: number; // 2nd Tarp [Entero]
    observations: string;
    created_at: string;
    updated_at: string;
}
export type OperationsOrlandoData = {
    id: number;
    closing_date: string;
    week: number;// Week [Entero]
    roof: number; // Roof [Entero]
    water: number; // Water [Entero]
    mold: number; // Mold [Entero]
    mold_testing: number; // Mold testing [Entero]
    shrink_wrap: number; // Shrink Wrap [Entero]
    inspections: number; // Inspections [Entero]
    second_tarp: number; // 2nd Tarp [Entero]
    observations: string;
    created_at: string;
    updated_at: string;
}
export type OperationsTexasData = {
    id: number;
    closing_date: string;
    week: number;// Week [Entero]
    roof: number; // Roof [Entero]
    inspections: number; // Inspections [Entero]
    second_tarp: number; // 2nd Tarp [Entero]
    observations: string;
    created_at: string;
    updated_at: string;
}